requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==1.1.1
boto3==1.0.0
flask-cors==3.0.8
mysql-connector-python==8.0.11